refactor: #13 - MemberInfoDto 클래스 data/memberInfo/dto 패키지로 위치 이전
