+++ backend/ezgg/src/main/java/com/matching/ezgg/matching/dto/MatchingSuccessResponse.java
@@ -1,6 +1,6 @@
 package com.matching.ezgg.matching.dto;
 
-import com.matching.ezgg.member.dto.MemberInfoDto;
+import com.matching.ezgg.data.memberInfo.dto.MemberInfoDto;
 import com.matching.ezgg.data.recentTwentyMatch.dto.RecentTwentyMatchDto;
 
 import lombok.AllArgsConstructor;
